remove cdc/owner/scheduler.go
